Move file attachment models into separate file
